rename typos.py -> typo.py
