+++ handlers/remote_handler.py
@@ -6,8 +6,8 @@ import shutil
 import time
 
 ## custom modules
-from entities.typos import typo as typo_blueprint
-from entities.typos import incorrectTypo as incorrect_typo_blueprint
+from entities.typo import typo as typo_blueprint
+from entities.typo import incorrectTypo as incorrect_typo_blueprint
 
 from entities.csep import csep as csep_blueprint
