+++ modules/score_rater.py
@@ -7,8 +7,8 @@ import msvcrt
 from entities.words import word
 from entities.vocab import vocab
 
-from entities.typos import typo as typo_blueprint
-from entities.typos import incorrectTypo as incorrect_typo_blueprint
+from entities.typo import typo as typo_blueprint
+from entities.typo import incorrectTypo as incorrect_typo_blueprint
 
 from modules.logger import Logger
 from modules.toolkit import Toolkit
